+++ app/controllers/admin/products_controller.rb
@@ -1,0 +1,37 @@
+class Admin::ProductsController < ApplicationController
+  def index
+    @products = Product.all
+  end
+
+  def new
+    @product = Product.new
+  end
+
+  def edit
+    @product = Product.find(params[:id])
+  end
+
+  def update
+    @product = Product.find(params[:id])
+    if @product.update(products_params)
+      redirect_to admin_products_path
+    else
+      render :eidt
+    end
+  end
+
+  def create
+    @product = Product.new(product_params)
+    if @product.save
+      redirect_to products_path
+    else
+      render :new
+    end
+  end
+
+  private
+
+  def product_params
+    params.require(:product).permit(:title, :description, :price, :quantity)
+  end
+end
