+++ app/helpers/admin/products_helper.rb
@@ -1,0 +1,2 @@
+module Admin::ProductsHelper
+end
