+++ app/views/admin/products/index.html.erb
@@ -1,0 +1,6 @@
+<ul>
+  <% @products.each do |product| %>
+
+    <li> <%= link_to(product.title, admin_product_path(product)) %> </li>
+  <% end %>
+</ul>
