+++ app/views/products/index.html.erb
@@ -1,4 +1,4 @@
-<div class="col-md-12">
+<!-- <div class="col-md-12">
   <div class="product">
     <%= link_to("New product", new_product_path, class: "btn btn-primary pull-right") %>
   </div>
@@ -29,4 +29,4 @@
       <% end %>
     </tbody>
   </table>
-</div>
+</div> -->
