+++ test/controllers/admin/products_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class Admin::ProductsControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
